+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [release candidate]
+
+### Added
+
+- Started renaming `set` to `set_value` and `get` to `get_value` ([#15])
+- Redis keys functions: `TYPE`, `SCAN`, `KEYS` ([#15])
+- Redis set functions: `SADD`, `SREM`, `SISMEMBER`, `SCARD`, `SMEMBERS` ([#15])
+
 ## [0.4.2] - 2023-11-19
 
 ### Bug-Fixes
@@ -23,7 +31,7 @@
 
 ### Added
 
-- Added ZRANGE (partially) and LRANGE. ([#7])
+- Added `ZRANGE` (partially) and `LRANGE`. ([#7])
 - Inferring backend in Redis constructor. ([#3])
 - Allow access to raw runtime and redis connections. ([#3])
 - Executing scripts from a pipeline. ([#3])
@@ -46,3 +54,4 @@
 
 [#3]: https://github.com/JosuaKrause/redipy/pull/3
 [#7]: https://github.com/JosuaKrause/redipy/pull/7
+[#15]: https://github.com/JosuaKrause/redipy/pull/15